moved filesystem setup steps from old makeImage.sh into _in_container_run.sh

+++ tests/unit_tests/_in_container_run.sh
@@ -29,7 +29,63 @@ md5sum /tmp/fs.tar.gz
 rm -rf "$d"
 
 # Now make us an image in /tmp/empty
-fakeroot /pkg/scripts/makeImage.sh $ARCH /tmp/fs.tar.gz /tmp/empty /pkg/resources /igloo_static
+#fakeroot /pkg/scripts/makeImage.sh $ARCH /tmp/fs.tar.gz /tmp/empty /pkg/resources /igloo_static
+
+#ARCH="$1"
+TARBALL="/tmp/fs.tar.gz"
+REPACK_DIR="/tmp/empty"
+RESOURCE_DIR="/pkg/resources"
+STATIC_DIR="/igloo_static"
+
+WORK_DIR="${REPACK_DIR}/work"
+TARFILE="${REPACK_DIR}/fs.tar"
+#IMAGE="${REPACK_DIR}/image.raw"
+#QCOW="${REPACK_DIR}/image.qcow"
+FD="${WORK_DIR}/firmadyne"
+IGLOO="${WORK_DIR}/igloo"
+#BLOCK_SIZE=4096
+
+# Validate and prepare tarball
+if [ ! -e "$TARBALL" ]; then
+    echo "Error: Cannot find tarball of root filesystem in $REPACK_DIR" >&2
+    exit 1
+fi
+
+# Create required directories
+mkdir -p "$WORK_DIR" "$FD" "$IGLOO/utils" "$IGLOO/keys"
+mkdir -p /tmp/artifacts
+
+# Extract fs.tar.gz -> work/fs.tar
+gunzip -c "$TARBALL" > "$TARFILE"
+
+# Populate work/* with the files we want to add in work/fd/ and work/igloo/
+# Copy keys and utilities
+cp "$RESOURCE_DIR/static_keys/"* "$IGLOO/keys"
+for t in "console" "libnvram" "utils.bin" "utils.source" "vpn"; do
+    UTILS="$STATIC_DIR/$t"
+    [ -d "$UTILS" ] || { echo "FATAL: Missing utilities directory $UTILS" >&2; exit 1; }
+
+    for f in "$UTILS"/*.${ARCH} "$UTILS"/*.all; do
+        [ -e "$f" ] || { echo "WARN: No files matching pattern $f"; continue; }
+        fname=$(basename "$f")
+        dest="${IGLOO}/utils/${fname%.*}"
+        cp "$f" "$dest"
+        chmod +x "$dest"
+    done
+done
+
+# Create symbolic links
+ln -s "/igloo/utils/busybox" "$IGLOO/utils/sh"
+ln -s "/igloo/utils/busybox" "$IGLOO/utils/sleep"
+
+# Validate and append this data to our tar file
+[ -e "$TARFILE" ] || { echo "Error: Tar file $TARFILE does not exist" >&2; exit 1; }
+[ -s "$TARFILE" ] || { echo "Error: Tar file $TARFILE is empty" >&2; exit 1; }
+tar --append --sort=name --owner=root:0 --group=root:0 --mtime='UTC 2019-01-01' -f "$TARFILE" -C "$WORK_DIR" .
+gzip "$TARFILE"
+
+fakeroot /pkg/scripts/makeImage.sh "${REPACK_DIR}/fs.tar.gz" /tmp/empty
+
 
 # Rewrite our config template to set the specified kernel version and arch values (in arch and kernel fields)
 cp /tests/$TESTDIR/config.yaml /tmp/config.yaml
